Remove @with_checker for threshold bork tests.
These tests used to sleep for up to 5 seconds, and they were executed
*many* times.  Tests for a single change to the VSC tools would
require more than 10 minutes to finish, with no new paths being
discovered.  The "fuzzed" parameters were just passed blindly to
sub-callees, and didn't affect the logic, the paths or the results of
the report_and_exit method in any way.

Now it takes only 2 seconds to run the Nagios tests, instead of > 10
minutes.